requirements.txt
# Transitive dependencies.
asn1crypto==0.24.0
cloudpickle==1.2.2
cryptography==2.1.4
future==0.18.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pycrypto==2.6.1
pyglet==1.3.2
pygobject==3.26.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.15.0

# Base dependencies.
absl-py==0.9.0
numpy==1.18.0
Pillow==7.1.2
python-snappy==0.5.4
scipy==1.2.1

# JAX libraries with GPU support.
jax==0.1.72
jaxlib @ https://storage.googleapis.com/jax-releases/cuda101/jaxlib-0.1.49-cp36-none-linux_x86_64.whl

# ALE and Gym libraries.
atari-py==0.2.6
gym==0.13.1

# DeepMind-specific libraries.
dm-env==1.2
dm-haiku @ git+git://github.com/deepmind/dm-haiku.git@db991d56563221d5a06be5e7228155e53d01aba9
dm-tree==0.1.5
rlax @ git+git://github.com/deepmind/rlax.git@870cba1ea8ad36725f4f3a790846298657b6fd4b
